This commit adds a hook to the Site Admin Updates page that will set versions.auto_upgrade true if readinessUpgrade detects no drift.
It is only triggered when the updates page is loaded and the readinessCheck query is triggered.

It may make more sense to run both readinessCheck and setAutoUpgrade queries as part of the updateCheck triggered as a routine.

This is still in drafting. The setAutoUpgrade query is also available to admins.