embeddings: add scoring function
This adds a scoring function which computes the weighted sum of
cosine similarity and document rank.

For now, both scores contribute equally.

Anecdotically, I saw in my experiments how a 'go.sum' file (with high
similarity) was ranked down because its low document rank.

I also added a new field "Debug" to the API which, similarily to Zoekt,
adds debug information about the score to the response if "Debug" is set
to true.

The API change is backward compatible.